Changes to support 64-bit build, python build materials now go in site_scons (sconsy way to do things)

+++ api/binding/value.cpp
@@ -365,7 +365,7 @@ namespace kroll
 		else if (this->IsVoidPtr())
 		{
 			void *value = this->ToVoidPtr();
-			oss << "<void* addr: " << (int) value << ">";
+			oss << "<void* addr: " << value << ">";
 		}
 		else if (this->IsNull())
 		{
